+++ sw-sysemu/conanfile.py
@@ -1,9 +1,9 @@
 from conan import ConanFile
-from conan.tools.cmake import CMakeToolchain, CMakeDeps, CMake
-from conan.tools.layout import cmake_layout
-from conans import tools
+from conan.tools.cmake import CMake, CMakeToolchain, CMakeDeps, cmake_layout
+from conan.tools.files import get, rmdir
 import os
-import re
+
+required_conan_version = ">=1.52.0"
 
 
 class SwSysemuConan(ConanFile):
@@ -29,7 +29,6 @@ class SwSysemuConan(ConanFile):
         "url": "git@gitlab.esperanto.ai:software/sw-sysemu.git",
         "revision": "auto",
     }
-    generators = "CMakeDeps"
 
     python_requires = "conan-common/[>=0.5.0 <1.0.0]"
 
@@ -51,7 +50,47 @@ class SwSysemuConan(ConanFile):
 
     def layout(self):
         cmake_layout(self)
-        self.cpp.source.includedirs = ["."]
+        self.cpp.source.includedirs = [".", "sw-sysemu/include"]
+
+        self.cpp.package.libs = ["fpu", "sysemu", "sw-sysemu"]
+        self.cpp.package.includedirs = ["include"]
+
+        libfpu_comp_name = "libfpu"
+        libfpu_cmake_name = "fpu"
+        self.cpp.package.components[libfpu_comp_name].set_property("cmake_target_name", libfpu_cmake_name)
+        self.cpp.package.components[libfpu_comp_name].requires = []
+        self.cpp.package.components[libfpu_comp_name].libs = ["fpu"]
+        self.cpp.package.components[libfpu_comp_name].includedirs =  ["include"]
+        self.cpp.build.components[libfpu_comp_name].libs = ["fpu"]
+        self.cpp.build.components[libfpu_comp_name].libdirs = ["."]
+        self.cpp.source.components[libfpu_comp_name].includedirs = [".", "fpu"]
+
+
+        libsysemu_comp_name = "libsysemu"
+        libsysemu_cmake_name = "sysemu"
+        self.cpp.package.components[libsysemu_comp_name].set_property("cmake_target_name", libsysemu_cmake_name)
+        self.cpp.package.components[libsysemu_comp_name].requires = ["elfio::elfio", libfpu_comp_name]
+        self.cpp.package.components[libsysemu_comp_name].libs = ["sysemu"]
+        self.cpp.package.components[libsysemu_comp_name].defines = ["SYS_EMU"]
+        self.cpp.package.components[libsysemu_comp_name].includedirs =  ["include"]
+        self.cpp.package.components[libsysemu_comp_name].libdirs = ["lib"]
+        if self.options.backtrace:
+            self.cpp.package.components[libsysemu_comp_name].requires.append("libunwind::libunwind")
+        self.cpp.build.components[libsysemu_comp_name].requires = ["elfio::elfio", libfpu_comp_name]
+        self.cpp.build.components[libsysemu_comp_name].libs = ["sysemu"]
+        self.cpp.build.components[libsysemu_comp_name].defines = ["SYS_EMU"]
+        self.cpp.build.components[libsysemu_comp_name].libdirs = ["."]
+        self.cpp.source.components[libsysemu_comp_name].includedirs = ["."]
+
+        libsw_sysemu_comp_name = "libsw-sysemu"
+        libsw_sysemu_cmake_name = "sw-sysemu"
+        self.cpp.package.components[libsw_sysemu_comp_name].set_property("cmake_target_name", libsw_sysemu_cmake_name)
+        self.cpp.package.components[libsw_sysemu_comp_name].requires = [libsysemu_comp_name, libfpu_comp_name, "glog::glog"]
+        self.cpp.package.components[libsw_sysemu_comp_name].libs = ["sw-sysemu"]
+        self.cpp.build.components[libsw_sysemu_comp_name].requires = [libsysemu_comp_name, libfpu_comp_name, "glog::glog"]
+        self.cpp.build.components[libsw_sysemu_comp_name].libs = ["sw-sysemu"]
+        self.cpp.build.components[libsw_sysemu_comp_name].libdirs = ["."]
+        self.cpp.source.components[libsw_sysemu_comp_name].includedirs = ["sw-sysemu/include", "sw-sysemu/include/sw-sysemu"]
 
     def generate(self):
         tc = CMakeToolchain(self)
@@ -59,6 +98,9 @@ class SwSysemuConan(ConanFile):
         tc.variables["BACKTRACE"] = self.options.backtrace
         tc.variables["CMAKE_INSTALL_LIBDIR"] = "lib"
         tc.generate()
+
+        deps = CMakeDeps(self)
+        deps.generate()
 
     def build(self):
         cmake = CMake(self)
@@ -68,50 +110,16 @@ class SwSysemuConan(ConanFile):
     def package(self):
         cmake = CMake(self)
         cmake.install()
-        tools.rmdir(os.path.join(self.package_folder, "lib", "cmake"))
+        rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
 
     def package_info(self):
-        # library components
-        libfpu_comp_name = "libfpu"
-        libfpu_cmake_name = "fpu"
-        self.cpp_info.components[libfpu_comp_name].names["cmake_find_package"] = libfpu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libfpu_comp_name].names["cmake_find_package_multi"] = libfpu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libfpu_comp_name].set_property("cmake_target_name", libfpu_cmake_name)
-        self.cpp_info.components[libfpu_comp_name].requires = []
-        self.cpp_info.components[libfpu_comp_name].libs = ["fpu"]
-        self.cpp_info.components[libfpu_comp_name].includedirs =  ["include"]
-        self.cpp_info.components[libfpu_comp_name].libdirs = ["lib"]
-
-        libsysemu_comp_name = "libsysemu"
-        libsysemu_cmake_name = "sysemu"
-        self.cpp_info.components[libsysemu_comp_name].names["cmake_find_package"] = libsysemu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libsysemu_comp_name].names["cmake_find_package_multi"] = libsysemu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libsysemu_comp_name].set_property("cmake_target_name", libsysemu_cmake_name)
-        self.cpp_info.components[libsysemu_comp_name].requires = ["elfio::elfio", libfpu_comp_name]
-        self.cpp_info.components[libsysemu_comp_name].libs = ["sysemu"]
-        self.cpp_info.components[libsysemu_comp_name].defines = ["SYS_EMU"]
-        self.cpp_info.components[libsysemu_comp_name].includedirs =  ["include"]
-        self.cpp_info.components[libsysemu_comp_name].libdirs = ["lib"]
-        if self.options.backtrace:
-            self.cpp_info.components[libsysemu_comp_name].requires.append("libunwind::libunwind")
-
-        libsw_sysemu_comp_name = "libsw-sysemu"
-        libsw_sysemu_cmake_name = "sw-sysemu"
-        self.cpp_info.components[libsw_sysemu_comp_name].names["cmake_find_package"] = libsw_sysemu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libsw_sysemu_comp_name].names["cmake_find_package_multi"] = libsw_sysemu_cmake_name # deprecated (remove with conan 2.0)
-        self.cpp_info.components[libsw_sysemu_comp_name].set_property("cmake_target_name", libsw_sysemu_cmake_name)
-        self.cpp_info.components[libsw_sysemu_comp_name].requires = [libsysemu_comp_name, libfpu_comp_name, "glog::glog"]
-        self.cpp_info.components[libsw_sysemu_comp_name].libs = ["sw-sysemu"]
-        self.cpp_info.components[libsw_sysemu_comp_name].includedirs =  ["include"]
-        self.cpp_info.components[libsw_sysemu_comp_name].libdirs = ["lib"]
-
         # utilities
-        bin_path = os.path.join(self.package_folder, "bin")
+        bin_path = os.path.join(self.package_folder, "bin") if self.package_folder else "bin"
         bin_ext = ".exe" if self.settings.os == "Windows" else ""
 
         self.output.info("Appending PATH env var with : {}".format(bin_path))
         self.env_info.PATH.append(bin_path)
 
-        sys_emu = os.path.join(self.package_folder, "bin", "sys_emu" + bin_ext)
+        sys_emu = os.path.join(bin_path, "sys_emu" + bin_ext)
         self.output.info("Setting SYS_EMU to {}".format(sys_emu))
         self.env_info.SYS_EMU = sys_emu